+++ common/src/process.rs
@@ -3,17 +3,17 @@ use clap_sys::process::*;
 
 #[repr(i32)]
 pub enum ProcessStatus {
-    Continue = CLAP_PROCESS_CONTINUE as i32,
-    ContinueIfNotQuiet = CLAP_PROCESS_CONTINUE_IF_NOT_QUIET as i32,
-    Sleep = CLAP_PROCESS_SLEEP as i32,
-    Tail = CLAP_PROCESS_TAIL as i32,
+    Continue = CLAP_PROCESS_CONTINUE,
+    ContinueIfNotQuiet = CLAP_PROCESS_CONTINUE_IF_NOT_QUIET,
+    Sleep = CLAP_PROCESS_SLEEP,
+    Tail = CLAP_PROCESS_TAIL,
 }
 
 impl ProcessStatus {
     pub fn from_raw(raw: clap_process_status) -> Option<Result<Self, ()>> {
         use ProcessStatus::*;
 
-        match raw as i32 {
+        match raw {
             CLAP_PROCESS_CONTINUE => Some(Ok(Continue)),
             CLAP_PROCESS_CONTINUE_IF_NOT_QUIET => Some(Ok(ContinueIfNotQuiet)),
             CLAP_PROCESS_SLEEP => Some(Ok(Sleep)),
